T4APartitionedMPSs

Documentation for T4APartitionedMPSs.

T4APartitionedMPSs.LazyContraction

T4APartitionedMPSs.PartitionedMPS

T4APartitionedMPSs.Projector

T4APartitionedMPSs.Projector

T4APartitionedMPSs.SubDomainMPS

Base.:&

Base.:+

Base.:<

Base.:|

Base.getindex

Base.isdisjoint

Base.keys

Base.length

Base.values

LinearAlgebra.norm

T4AITensorCompat.contract

T4AITensorCompat.contract

T4AITensorCompat.truncate

T4APartitionedMPSs._add_patching

T4APartitionedMPSs._next_projindex

T4APartitionedMPSs.adaptive_patching

T4APartitionedMPSs.adaptive_patching

T4APartitionedMPSs.adaptivecontract

T4APartitionedMPSs.add_patching

T4APartitionedMPSs.automul

T4APartitionedMPSs.combinesites

T4APartitionedMPSs.contract!

T4APartitionedMPSs.elemmul

T4APartitionedMPSs.extractdiagonal

T4APartitionedMPSs.findallsites_by_tag

T4APartitionedMPSs.makesitediagonal

T4APartitionedMPSs.noprime

T4APartitionedMPSs.prime

T4APartitionedMPSs.projcontract

T4APartitionedMPSs.projcontract

T4APartitionedMPSs.project

T4APartitionedMPSs.rearrange_siteinds

T4APartitionedMPSs.siteindices

T4APartitionedMPSs.LazyContraction — Type

Lazy evaluation for contraction of two SubDomainMPS objects.

T4APartitionedMPSs.PartitionedMPS — Type

PartitionedMPS is a structure that holds multiple MPSs (SubDomainMPS) that are associated with different non-overlapping projectors.

T4APartitionedMPSs.Projector — Type

A projector represents a projection of a tensor from a set of its indices to integers. Each index is projected to a positive integer.

T4APartitionedMPSs.Projector — Method

Constructing a projector from a single pair of index and integer.

T4APartitionedMPSs.SubDomainMPS — Type

An MPS with a projector.

Base.:& — Method

a & b represents the intersection of the indices that a and b are projected at.

Base.:+ — Method

Add two PartitionedMPS objects.

If the two projects have the same projectors in the same order, the resulting PartitionedMPS will have the same projectors in the same order. By default, we use directsum algorithm to compute the sum and no truncation is performed.

Base.:< — Method

a < b means that a is projected at a subset of the indices that b is projected at.

Base.:| — Method

a | b represents the union of the indices that a and b are projected at.

If a is projected at inds=1 and b is not projected for the same inds, then a | b is not projected for inds.

Base.getindex — Method

Indexing for PartitionedMPS. This is deprecated and will be removed in the future.

Base.isdisjoint — Method

Return if projectors are not overlapping

Base.keys — Method

Return the keys, i.e., projectors of the PartitionedMPS.

Base.length — Method

Get the number of the data in the PartitionedMPS. This is NOT the number of sites in the PartitionedMPS.

Base.values — Method

Return the values, i.e., SubDomainMPS of the PartitionedMPS.

LinearAlgebra.norm — Method

Return the norm of the PartitionedMPS.

T4AITensorCompat.contract — Method

Contract two T4APartitionedMPSs MPS objects.

At each site, the objects must share at least one site index.

T4AITensorCompat.contract — Method

Contraction of two SubDomainMPSs. Only if the shared projected indices overlap the contraction is non-vanishing.

T4AITensorCompat.truncate — Method

Truncate a PartitionedMPS object piecewise.

Each SubDomainMPS in the PartitionedMPS is truncated independently, but the cutoff is adjusted according to the norm of each SubDomainMPS. The total error is the sum of the errors in each SubDomainMPS.

T4APartitionedMPSs._add_patching — Method

Add multiple SubDomainMPS objects on the same projector.

If the bond dimension of the result reaches maxdim, perform patching recursively to reduce the bond dimension.

T4APartitionedMPSs._next_projindex — Method

Return the next index to be projected.

T4APartitionedMPSs.adaptive_patching — Method

Adaptive patching

Do patching recursively to reduce the bond dimension. If the bond dimension of a SubDomainMPS exceeds maxdim, perform patching.

T4APartitionedMPSs.adaptive_patching — Method

Adaptive patching

Do patching recursively to reduce the bond dimension. If the bond dimension of a SubDomainMPS exceeds maxdim, perform patching.

T4APartitionedMPSs.adaptivecontract — Method

Perform contraction of two PartitionedMPS objects.

The resulting patches after the contraction are patch-added if projected on the same final patch.

T4APartitionedMPSs.add_patching — Method

Add multiple PartitionedMPS objects.

T4APartitionedMPSs.automul — Method

function automul(
    M1::PartitionedMPS,
    M2::PartitionedMPS;
    tag_row::String="",
    tag_shared::String="",
    tag_col::String="",
    ...

)

Performs automatic multiplication between partitioned MPSs. Automatic multiplication is defined as:

\[ (fg) (\sigma_{row}, \sigma_{col}; \xi) = \sum_{\sigma_{shared}} f(\sigma_{row}, \sigma_{shared}; \xi) g(\sigma_{shared}, \sigma_{col} ; \xi).\]

By default, only element-wise product on sites $\xi$ will be performed. See also: elemmul.

T4APartitionedMPSs.combinesites — Method

Contract two adjacent tensors in TensorTrain

T4APartitionedMPSs.contract! — Method

Contract two PartitionedMPS objects.

Existing patches M in the resulting PartitionedMPS will be overwritten if overwrite=true.

T4APartitionedMPSs.elemmul — Method

function elemmul(
    M1::PartitionedMPS,
    M2::PartitionedMPS
)

Performs elementwise multiplication between partitioned MPSs. Element-wise product is defined as:

\[ (fg) (\xi) = f(\xi)g(\xi) = \sum_{\xi'} f(\xi, \xi') g(\xi')\]

where $f(\xi, \xi') = f(\xi) \delta_{\xi, \xi'}$.

T4APartitionedMPSs.extractdiagonal — Method

Extract diagonal of the PartitionedMPS for s, s', ... for a given site index s, where s must have a prime level of 0.

T4APartitionedMPSs.findallsites_by_tag — Method

Find sites with the given tag

For tag = x, if sites contains an Index object with x, the function returns a vector containing only its positon.

If not, the function seach for all Index objects with tags x=1, x=2, ..., and return their positions.

If no Index object is found, an empty vector will be returned.

T4APartitionedMPSs.makesitediagonal — Method

Make the PartitionedMPS diagonal for a given site index s by introducing a dummy index s'.

T4APartitionedMPSs.noprime — Method

Remove prime level in projected indices.

T4APartitionedMPSs.prime — Function

Prime projected indices.

T4APartitionedMPSs.projcontract — Method

Project SubDomainMPS vectors to proj before computing all possible pairwise contractions of the elements. The results are summed or patch-summed if belonging to the same patch.

T4APartitionedMPSs.projcontract — Method

Project two SubDomainMPS objects to proj before contracting them.

T4APartitionedMPSs.project — Method

Project the LazyContraction object to prj before evaluating it.

This may result in projecting the external indices of a and b.

T4APartitionedMPSs.rearrange_siteinds — Method

Rearrange the site indices of the PartitionedMPS according to the given order. If nessecary, tensors are fused or split to match the new order.

T4APartitionedMPSs.siteindices — Method

Return the site indices of the PartitionedMPS. The site indices are returned as a vector of sets, where each set corresponds to the site indices at each site.